+++ mininterface/textual_interface/textual_app.py
@@ -1,7 +1,9 @@
 from typing import TYPE_CHECKING
 from textual import events
 from textual.app import App, ComposeResult
-from textual.containers import VerticalScroll
+from textual.content import Content
+from textual.binding import Binding
+from textual.containers import VerticalScroll, Container
 from textual.widget import Widget
 from textual.widgets import (
     Checkbox,
@@ -56,13 +58,9 @@ class TextualApp(App[bool | None]):
 
     def __init__(self, adaptor: "TextualAdaptor", submit: str | bool = True):
         super().__init__()
-        self.title = adaptor.facet._title
         self.widgets: WidgetList = []
-        self.focusable: WidgetList = []
-        """ A subset of self.widgets"""
-        self.focused_i: int = 0
+        self.focusable_: WidgetList = []
         self.adaptor = adaptor
-        self.output = Static("")
         self.submit = submit
 
         # Form confirmation
@@ -71,10 +69,43 @@ class TextualApp(App[bool | None]):
             self.bind("Enter", "confirm", description=submit if isinstance(submit, str) else "Ok")
         self.bind("escape", "exit", description="Cancel")
 
+    def compose(self):
+        self.contents = Container()
+        yield self.contents
+
+    def on_mount(self):
+        self.contents.remove_children()
+        self.contents.mount(MainContents(self.adaptor, self.widgets, self.focusable_))
+
+    def action_confirm(self):
+        # next time, start on the same widget
+        # NOTE the functionality is probably not used
+        self.focused_i = next((i for i, inp in enumerate(self.focusable_) if inp == self.focused), None)
+        self.exit(True)
+
+    def action_exit(self):
+        self.exit()
+
+
+class MainContents(Static):
+
+    def __init__(self, adaptor: "TextualAdaptor", widgets: WidgetList, focusable_: WidgetList):
+        super().__init__()
+        self.app: "TextualApp"
+        self.title = adaptor.facet._title  # TODO – rather here?
+        self.widgets = widgets
+        self.focusable_ = focusable_
+        """ A subset of self.widgets"""
+        self.focused_i: int = 0
+        self.adaptor = adaptor
+        self.output = Static("")
+
     def compose(self) -> ComposeResult:
         # prepare widgets
         # since textual 1.0.0 we have to build widgets not earlier than the context app is ready
-        self.widgets = list(flatten(formdict_to_widgetdict(
+
+        self.widgets.clear()
+        self.widgets.extend(flatten(formdict_to_widgetdict(
             self.adaptor.facet._form, self.adaptor.widgetize), include_keys=self.adaptor.header))
 
         # there are multiple sections in the list, <hr>ed by Rule elements. However, the first takes much space.
@@ -94,6 +125,8 @@ class TextualApp(App[bool | None]):
             for i, fieldt in enumerate(self.widgets):
                 if isinstance(fieldt, Input):
                     yield Label(fieldt.placeholder)
+                # TODO MyRadioSet not shown now: add name in widgetize and display here
+                # NOTE: has this something to do with the PathTag?
                 elif hasattr(fieldt, "_link") and fieldt._link.name and not isinstance(fieldt, Input):
                     yield Label(fieldt._link.name)
                 yield fieldt
@@ -105,37 +138,32 @@ class TextualApp(App[bool | None]):
                         self.focused_i = i
                     yield Label(desc)
                 yield Label("")
-        self.focusable = [w for w in self.widgets if isinstance(w, (Input, Changeable))]
+        self.focusable_.clear()
+        self.focusable_.extend(w for w in self.widgets if isinstance(w, (Input, Changeable)))
 
     def on_mount(self):
         self.widgets[self.focused_i].focus()
 
-    def action_confirm(self):
-        # next time, start on the same widget
-        # NOTE the functionality is probably not used
-        self.focused_i = next((i for i, inp in enumerate(self.focusable) if inp == self.focused), None)
-        self.exit(True)
-
-    def action_exit(self):
-        self.exit()
-
     def on_key(self, event: events.Key) -> None:
+        f = self.focusable_
         try:
-            index = self.focusable.index(self.focused)
+            index = f.index(self.app.focused)
         except ValueError:  # probably some other element were focused
             return
         match event.key:
             case "down":
-                self.focusable[(index + 1) % len(self.focusable)].focus()
+                f[(index + 1) % len(f)].focus()
+                event.stop()
             case "up":
-                self.focusable[(index - 1) % len(self.focusable)].focus()
+                f[(index - 1) % len(f)].focus()
+                event.stop()
             case "enter":
                 # NOTE a multiline input might be
                 # isinstance(self.focused,
-                if self.submit:
-                    self.action_confirm()
+                if self.app.submit:
+                    self.app.action_confirm()
             case letter if len(letter) == 1:  # navigate by letters
-                for inp_ in self.focusable[index+1:] + self.focusable[:index]:
+                for inp_ in f[index+1:] + f[:index]:
                     match inp_:
                         case Checkbox():
                             label = inp_.label